Move InfoPanelController to kit module
As all other panel controllers reside in the kit module, there is no
reason why the InfoPanelController shouldn't be there, as well.
