+++ app/src/main/resources/com/oracle/javafx/scenebuilder/app/i18n/SceneBuilderApp.properties
@@ -364,19 +364,6 @@ message.bar.details = Click to access the list of warning messages
 message.panel.clear = Clear
 
 # -----------------------------------------------------------------------------
-# Info panel
-# -----------------------------------------------------------------------------
-info.label.assigned = Assigned fx:id
-info.label.component = Component
-info.label.use.fx.root = Use fx:root construct
-# Used after number 1
-info.label.item = item
-# Used after any number greater than 1
-info.label.items = items
-# The parameter is the name of a Java class
-info.tooltip.controller = Your Controller class must extend {0}
-
-# -----------------------------------------------------------------------------
 # Menu Bar
 # -----------------------------------------------------------------------------
 menubar.no.lib.item = No Library Items
