+++ kit/src/main/resources/com/oracle/javafx/scenebuilder/kit/editor/panel/info/InfoPanel.fxml
