+++ kit/src/main/resources/com/oracle/javafx/scenebuilder/kit/i18n/SceneBuilderKit.properties
@@ -100,6 +100,19 @@ hierarchy.unresolved.resource = Unresolved resource
 hierarchy.unsupported.expression = Unsupported by Scene Builder
 
 # -----------------------------------------------------------------------------
+# Info panel
+# -----------------------------------------------------------------------------
+info.label.assigned = Assigned fx:id
+info.label.component = Component
+info.label.use.fx.root = Use fx:root construct
+# Used after number 1
+info.label.item = item
+# Used after any number greater than 1
+info.label.items = items
+# The parameter is the name of a Java class
+info.tooltip.controller = Your Controller class must extend {0}
+
+# -----------------------------------------------------------------------------
 # Content Panel
 # -----------------------------------------------------------------------------
 content.label.status.cannot.display = Content of this document cannot be displayed
